+++ src/deployconfig.jl
@@ -73,7 +73,7 @@ function deploy_folder(cfg::DeployConfig; kwargs...)
     return DeployDecision(; all_ok = false)
 end
 function deploy_folder(::Nothing; kwargs...)
-    @warn "Documenter could not auto-detect the building environment Skipping deployment."
+    @warn "Documenter could not auto-detect the building environment. Skipping deployment."
     return DeployDecision(; all_ok = false)
 end
 
@@ -934,9 +934,9 @@ pipeline:
 ```
 
 More about pipeline syntax is documented here: 
-- 0.15.x <https://woodpecker-ci.org/docs/0.15/usage/pipeline-syntax>
-- 1.0.0 and onwards <https://woodpecker-ci.org/docs/1.0/usage/pipeline-syntax>
-- 2.0.0 <https://woodpecker-ci.org/docs/usage/workflow-syntax>
+- 0.15.x: [https://woodpecker-ci.org/docs/0.15/usage/pipeline-syntax (hosted at archive.org; the documentation is no longer available on the Woodpecker website)](https://web.archive.org/web/20240318223506/https://woodpecker-ci.org/docs/0.15/usage/pipeline-syntax)
+- 1.0.0 and onwards: <https://woodpecker-ci.org/docs/1.0/usage/pipeline-syntax>
+- 2.0.0: <https://woodpecker-ci.org/docs/usage/workflow-syntax>
 """
 struct Woodpecker <: DeployConfig
     woodpecker_ci_version::VersionNumber